Fix static-exchange-evaluation implementation and add tests
Bench: 1452718
